Feedback from code reeview
